+++ src/lib.rs
@@ -37,7 +37,7 @@ pub use self::collections::Indexable;
 pub(crate) use self::collections::{IndexedMap, IndexedMapUpdate, Map};
 pub use self::column_statistics::*;
 pub use self::event::{Event, EventDb, EventKind, EventMessage, ThreatLevel};
-pub use self::migration::migrate_data_dir;
+pub use self::migration::{migrate_data_dir, write_version_markers};
 pub use self::model::{Digest, Model};
 pub use self::scores::Scores;
 use self::tables::StateDb;